+++ @angular-generic-table/core/components/generic-table.component.ts
@@ -51,7 +51,7 @@ import { GtMetaPipe } from '../pipes/gt-meta.pipe';
                       (keyup)="gtColumnSearch(column.objectKey, columnSearch.value, false)"
                       type="text"
                       placeholder="Filter by {{gtFields | gtProperty:column.objectKey:'name' | lowercase}}" />
-                    <gt-checkbox *ngIf="(gtFields | gtProperty:column.objectKey:'columnComponent')?.type === 'checkbox'" [checked]="(selectedRows.length === gtData.length)" (changed)="(selectedRows.length !== gtData.length) ? selectAllRows() : deselectAllRows();"></gt-checkbox>
+                    <gt-checkbox *ngIf="(gtFields | gtProperty:column.objectKey:'columnComponent')?.type === 'checkbox'" [checked]="(selectedRows.length === gtData.length)" (changed)="toggleAllRows()"></gt-checkbox>
                 </th>
             </tr>
             </thead>
@@ -532,8 +532,8 @@ export class GenericTableComponent<R extends GtRow, C extends GtExpandedRow<R>>
 					}
 					break;
 			}
-		} /* if ctrl key or meta key is not press together with sort... */ else {
-			switch (pos) {
+		} else {
+			/* if ctrl key or meta key is not press together with sort... */ switch (pos) {
 				// ...and property is not sorted before...
 				case -1:
 					// ...sort by property
@@ -795,6 +795,16 @@ export class GenericTableComponent<R extends GtRow, C extends GtExpandedRow<R>>
 	public deselectAllRows(): void {
 		this._toggleAllRowProperty('isSelected', false);
 	}
+	/**
+	 * Toggle all rows.
+	 */
+	public toggleAllRows(): void {
+		if (this.selectedRows.length !== this.gtData.length) {
+			this.selectAllRows();
+		} else {
+			this.deselectAllRows();
+		}
+	}
 
 	/**
 	 * Toggle row collapsed state ie. expanded/open or collapsed/closed.
@@ -892,7 +902,10 @@ export class GenericTableComponent<R extends GtRow, C extends GtExpandedRow<R>>
 		source: Array<GtRow>
 	): Array<GtRow> {
 		for (let i = 0; i < source.length; i++) {
-			target.push(source[i]);
+			// only add if not already in list
+			if (target.indexOf(source[i]) === -1) {
+				target.push(source[i]);
+			}
 		}
 		return target;
 	}
@@ -1543,10 +1556,8 @@ export class GenericTableComponent<R extends GtRow, C extends GtExpandedRow<R>>
 				if (setting.sort === 'asc') {
 					// ... add to sorting
 					sorting.push(setting.objectKey);
-				} /* ...else if sorted descending... */ else if (
-					setting.sort === 'desc'
-				) {
-					// ... add to sorting
+				} else if (setting.sort === 'desc') {
+					/* ...else if sorted descending... */ // ... add to sorting
 					sorting.push('-' + setting.objectKey);
 				}
 			}